Split work-in-progress new backend into its own files
